fix: stepped list width for empty lists
The width of lists is fixed and stepped with media queries, calculated
to fit the max number of columns based on the screen width. The widths
are calculated with `min-width` and there is no query for less than one
column. This caused empty lists to shrink due them having no children to
enforce the width.